+++ packages/react/src/types/templates.ts
@@ -1,5 +1,4 @@
 import {
-    KPIToken,
     ResolvedKPITokenWithData,
     ResolvedOracleWithData,
 } from "@carrot-kpi/sdk";
@@ -55,6 +54,14 @@ export interface BaseRemoteTemplateComponentProps {
     t: NamespacedTranslateFunction;
 }
 
+export type TemplateComponentStateUpdater<S extends SerializableObject<S>> = (
+    previousState: S,
+) => S;
+
+export type TemplateComponentStateChangeCallback<
+    S extends SerializableObject<S>,
+> = (stateOrUpdater: S | TemplateComponentStateUpdater<S>) => void;
+
 export interface OracleInitializationBundle {
     data: Hex;
     value: bigint;
@@ -63,17 +70,21 @@ export interface OracleInitializationBundle {
 export type OracleInitializationBundleGetter =
     () => Promise<OracleInitializationBundle>;
 
-export type OracleChangeCallback<S extends SerializableObject<S>> = (
-    internalState: S,
+export type OracleInitializationBundleGetterChangeCallback = (
     initializationBundleGetter?: OracleInitializationBundleGetter,
+) => void;
+
+export type OracleSuggestedExpirationTimestampChangeCallback = (
+    suggestedExpirationDate?: number,
 ) => void;
 
 export type AdditionalRemoteOracleCreationFormProps<
     S extends SerializableObject<S>,
 > = {
     state: S;
-    kpiToken?: Partial<KPIToken>;
-    onChange: OracleChangeCallback<S>;
+    onStateChange: TemplateComponentStateChangeCallback<S>;
+    onInitializationBundleGetterChange: OracleInitializationBundleGetterChangeCallback;
+    onSuggestedExpirationTimestampChange: OracleSuggestedExpirationTimestampChangeCallback;
     navigate: NavigateFunction;
     onTx: <T extends TxType>(tx: Tx<T>) => void;
 };
@@ -82,10 +93,6 @@ export type OracleRemoteCreationFormProps<S extends SerializableObject<S>> =
     BaseRemoteTemplateComponentProps &
         AdditionalRemoteOracleCreationFormProps<S>;
 
-export type KPITokenChangeCallback<S extends SerializableObject<S>> = (
-    state: S,
-) => void;
-
 export type OracleCreationFormProps<S extends SerializableObject<S>> =
     TemplateComponentProps & AdditionalRemoteOracleCreationFormProps<S>;
 
@@ -93,7 +100,7 @@ export type AdditionalRemoteKPITokenCreationFormProps<
     S extends SerializableObject<S>,
 > = {
     state: S;
-    onChange: KPITokenChangeCallback<S>;
+    onStateChange: TemplateComponentStateChangeCallback<S>;
     onCreate: () => void;
     navigate: NavigateFunction;
     onTx: <T extends TxType>(tx: Tx<T>) => void;